Update arcapi.py
First off, this is awesome.  I saw that in your "distinct" function you had an optional parameter for a where clause (w), and the w option was not in the arguments for the "values" function.

I also added a couple functions (nothing too fancy) I use all the time at the bottom for raster processing.  There are a couple for generating remap ranges for both 3d analyst and spatial analyst and one for finding certain file types.

I then added a couple tools to fill NoData values in a raster, convert integer raster to float and to convert DEM's whose elevation values are in meters to feet.  Feel free to reject any of the stuff I added.  I think this open source arcpy is a great idea and I hope to contribute more to it in the future!

-Caleb Mackey